// src/itunes/interfaces/itunes-response.interface.ts
export interface ItunesSearchResponse {
    resultCount: number;
    results: ItunesResult[];
  }
  
  export interface ItunesResult {
    wrapperType?: string;
    kind?: string;
    trackId: number;
    artistName: string;
    collectionName?: string;
    trackName: string;
    collectionCensoredName?: string;
    trackCensoredName?: string;
    collectionViewUrl?: string;
    feedUrl?: string;
    trackViewUrl: string;
    artworkUrl30?: string;
    artworkUrl60?: string;
    artworkUrl100?: string;
    collectionPrice?: number;
    trackPrice?: number;
    collectionHdPrice?: number;
    releaseDate?: string;
    collectionExplicitness?: string;
    trackExplicitness?: string;
    trackCount?: number;
    trackTimeMillis?: number;
    country: string;
    currency?: string;
    primaryGenreName?: string;
    contentAdvisoryRating?: string;
    artworkUrl600?: string;
    genreIds?: string[];
    genres?: string[];
  }
  